Enhance winrm enable feature ensure winrm port is open
This change enhances the previous added winrm enable at launch feature that it
ensures that the winrm port is opened in the security group. If it is not, it
will notify user and add the winrm port to the current security group for the
new launched instance so that the winrm port is gaurentee opened at launch.